+++ src/main/java/com/wenxin2/marioverse/datagen/ItemModelGen.java
@@ -122,7 +122,6 @@ public class ItemModelGen extends ItemModelProvider {
         this.basicItem(ItemRegistry.WHITE_KOOPA_SHOES.get());
 
         this.handheldItem(BlockRegistry.CLASSIC_CHECKPOINT_FLAG.asItem());
-        this.handheldItem(BlockRegistry.CLASSIC_GOAL_POLE.asItem());
         this.handheldItem(ItemRegistry.CREATIVE_WRENCH.get());
         this.handheldItem(ItemRegistry.WRENCH.get());
         this.handheldItem(ItemRegistry.WARP_DISRUPTOR.get());
@@ -131,14 +130,40 @@ public class ItemModelGen extends ItemModelProvider {
         this.plasticFluidBucketItem(ItemRegistry.PLASTIC_WATER_BUCKET.get(), Fluids.WATER, true, false);
 
         for (Map.Entry<DyeColor, DeferredBlock<Block>> entry : BlockRegistry.CHECKPOINT_FLAGS.entrySet())
-            this.handheldItem(entry.getValue().asItem());
+            this.handheldItem(entry.getValue().asItem()).override()
+                    .model(new ModelFile.UncheckedModelFile(modLoc("item/american_checkpoint_flag")))
+                    .predicate(modLoc("custom_name"), 1.0F).end().override()
+                    .model(new ModelFile.UncheckedModelFile(modLoc("item/wonder_checkpoint_flag")))
+                    .predicate(modLoc("custom_name"), 2.0F).end();
+        this.getBuilder("american_checkpoint_flag").parent(new ModelFile.UncheckedModelFile("item/handheld"))
+                .texture("layer0", ResourceLocation.fromNamespaceAndPath(Marioverse.MOD_ID, "item/american_checkpoint_flag"));
+        this.getBuilder("wonder_checkpoint_flag").parent(new ModelFile.UncheckedModelFile("item/handheld"))
+                .texture("layer0", ResourceLocation.fromNamespaceAndPath(Marioverse.MOD_ID, "item/wonder_checkpoint_flag"));
 
         for (Map.Entry<DyeColor, DeferredBlock<Block>> entry : BlockRegistry.GOAL_POLES.entrySet())
-            this.handheldItem(entry.getValue().asItem());
+            this.handheldItem(entry.getValue().asItem()).override()
+                    .model(new ModelFile.UncheckedModelFile(modLoc("item/american_goal_pole")))
+                    .predicate(modLoc("custom_name"), 1.0F).end().override()
+                    .model(new ModelFile.UncheckedModelFile(modLoc("item/wonder_goal_pole")))
+                    .predicate(modLoc("custom_name"), 2.0F).end();
+        this.getBuilder("american_goal_pole").parent(new ModelFile.UncheckedModelFile("item/handheld"))
+                .texture("layer0", ResourceLocation.fromNamespaceAndPath(Marioverse.MOD_ID, "item/american_goal_pole"));
+        this.getBuilder("wonder_goal_pole").parent(new ModelFile.UncheckedModelFile("item/handheld"))
+                .texture("layer0", ResourceLocation.fromNamespaceAndPath(Marioverse.MOD_ID, "item/wonder_goal_pole"));
+
+        this.handheldItem(BlockRegistry.CLASSIC_GOAL_POLE.asItem()).override()
+                .model(new ModelFile.UncheckedModelFile(modLoc("item/american_classic_goal_pole")))
+                .predicate(modLoc("custom_name"), 1.0F).end().override()
+                .model(new ModelFile.UncheckedModelFile(modLoc("item/wonder_classic_goal_pole")))
+                .predicate(modLoc("custom_name"), 2.0F).end();
+        this.getBuilder("american_classic_goal_pole").parent(new ModelFile.UncheckedModelFile("item/handheld"))
+                .texture("layer0", ResourceLocation.fromNamespaceAndPath(Marioverse.MOD_ID, "item/american_classic_goal_pole"));
+        this.getBuilder("wonder_classic_goal_pole").parent(new ModelFile.UncheckedModelFile("item/handheld"))
+                .texture("layer0", ResourceLocation.fromNamespaceAndPath(Marioverse.MOD_ID, "item/wonder_classic_goal_pole"));
 
         this.basicItem(ItemRegistry.GOOMBA_SPAWN_EGG.get()).override()
-                .model(new ModelFile.UncheckedModelFile(ResourceLocation.fromNamespaceAndPath(Marioverse.MOD_ID, "item/goombella_spawn_egg")))
-                .predicate(ResourceLocation.fromNamespaceAndPath(Marioverse.MOD_ID, "goombella"), 1.0F).end();
+                .model(new ModelFile.UncheckedModelFile(modLoc("item/goombella_spawn_egg")))
+                .predicate(modLoc("custom_name"), 1.0F).end();
         this.getBuilder("goombella_spawn_egg").parent(new ModelFile.UncheckedModelFile("item/generated"))
                 .texture("layer0", ResourceLocation.fromNamespaceAndPath(Marioverse.MOD_ID, "item/goombella_spawn_egg"));
     }